// packages/test/generatedCode/education/classes/item/educationClassItemRequestBuilderGetRequestConfiguration.ts
import {EducationClassItemRequestBuilderGetQueryParameters} from './educationClassItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface EducationClassItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: EducationClassItemRequestBuilderGetQueryParameters;
}